Change mechanism of play management plugin
