+++ readme.md
@@ -226,7 +226,7 @@ Configure your dependencies
 ---------------------------
 
     resolvers += "Guardian Github Snapshots" at "http://guardian.github.com/maven/repo-releases"
-    libraryDependencies += "com.gu" %% "management-play" % "5.10"
+    libraryDependencies += "com.gu" %% "management-play" % "5.11"
 
 Add to the play plugins file
 ----------------------------
@@ -235,21 +235,19 @@ Add the following line to conf/play.plugins (create it if it doesn't exist):
 
     1000:com.gu.management.play.InternalManagementPlugin
 
-Add your application name
--------------------------
-
-Add your application name to your application.conf:
-
-    application.name="my-application-name"
-
 Bind the management pages
 -------------------------
 
-Create a scala Object somewhere in your project that binds in your list of management pages:
+The plugin locates the pages and application name by convention.
+
+Create a scala Object called conf.Management that mixes in the ManagementPageManifest trait
+binds in your list of management pages:
 
 ```scala
-object Management {
-  val applicationName = app.configuration.getString("application.name").get
+package conf
+
+object Management extends ManagementPageManifest {
+  val applicationName = "your-application-name"
 
   lazy val pages = List(
     new ManifestPage,
@@ -258,11 +256,11 @@ object Management {
     StatusPage(applicationName, Metrics.all),
     new LogbackLevelPage(applicationName)
   )
-
-  val plugin = Play.current.plugin[InternalManagementPlugin]
-  plugin.foreach { _.registerPages(pages) }
 }
 ```
+
+If there is a reason you can't name your file conf.Management, you can override it
+by setting management.manifestobject in application.conf.
 
 
 Providing metrics for GANGLIA
